refactor CartDetails and improve CarLineItem styling
