+++ templates/demo-store-neue/src/components/cart/index.js
@@ -1,1 +1,3 @@
 export {CartDetails} from './CartDetails.client';
+export {CartEmpty} from './CartEmpty.client';
+export {CartLineItem} from './CartLineItem.client';
